fixed duplicate video preset id
